n8n automation - antigravity (DND)/requirements.txt
# Core dependencies for the 3-layer architecture

# Environment management
python-dotenv==1.0.0

# HTTP requests
requests==2.31.0

# Data processing
pandas==2.1.4

# Notion API
notion-client==2.2.1

# Utilities
python-dateutil==2.8.2
